requirements.txt
feedparser>=6.0.11
requests>=2.31.0
pyyaml>=6.0.1
trafilatura>=1.8.0
openai>=1.40.0
python-dateutil>=2.9.0.post0
